Move Geometry.Changed handler to Shape class to make it available to all inheritors
